flask-app/pyproject.toml
[project]
name = "flask-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.0",
    "flask>=3.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
